found issue in migration create_permission_table

+++ database/migrations/2018_01_01_000000_create_permission_tables.php
@@ -30,7 +30,7 @@ class CreatePermissionTables extends Migration
         });
 
         Schema::create($tableNames['model_has_permissions'], function (Blueprint $table) use ($tableNames) {
-            $table->increments('id');
+            $table->bigInteger('id');
             $table->unsignedInteger('permission_id');
             $table->morphs('model');
 
@@ -43,7 +43,7 @@ class CreatePermissionTables extends Migration
         });
 
         Schema::create($tableNames['model_has_roles'], function (Blueprint $table) use ($tableNames) {
-            $table->increments('id');
+            $table->bigInteger('id');
             $table->unsignedInteger('role_id');
             $table->morphs('model');
 
@@ -56,7 +56,7 @@ class CreatePermissionTables extends Migration
         });
 
         Schema::create($tableNames['role_has_permissions'], function (Blueprint $table) use ($tableNames) {
-            $table->increments('id');
+            $table->bigInteger('id');
             $table->unsignedInteger('permission_id');
             $table->unsignedInteger('role_id');
